+++ projects/journey-maps-client/src/lib/journey-maps-client.component.ts
@@ -82,9 +82,9 @@ export class JourneyMapsClientComponent implements OnInit, AfterViewInit, OnDest
    */
   @Output() mapCenterChange = new EventEmitter<LngLatLike>();
   /**
-   * This event is emitted whenever a marker, with property emitOnSelect, is selected.
+   * This event is emitted whenever a marker, with property emitOnSelect, is selected or unselected.
    */
-  @Output() selectedMarkerEmitter = new EventEmitter<Marker>();
+  @Output() selectedMarkerIdChange = new EventEmitter<string>();
 
   private mapCenterChangeDebouncer = new Subject<void>();
   private windowResized = new Subject<void>();
@@ -157,7 +157,9 @@ export class JourneyMapsClientComponent implements OnInit, AfterViewInit, OnDest
 
   public set selectedMarker(value: Marker) {
     if (value && value.emitOnSelect) {
-      this.selectedMarkerEmitter.emit(value);
+      this.selectedMarkerIdChange.emit(value.id);
+    } else {
+      this.selectedMarkerIdChange.emit(undefined);
     }
     if (value && value.markerUrl) {
       open(value.markerUrl, '_self'); // Do we need to make target configurable ?
@@ -260,6 +262,25 @@ export class JourneyMapsClientComponent implements OnInit, AfterViewInit, OnDest
     if (this.map?.isStyleLoaded() && value) {
       this.mapParameterChanged.next();
     }
+  }
+
+  /**
+   * If providing an ID, a marker with this ID must be present in the list of {@link markers}
+   *
+   * @param value the ID of the marker to select or <code>undefined</code> to unselect the marker
+   */
+  @Input()
+  set selectedMarkerId(value: string) {
+    if (!!value) {
+      const selectedMarker = this.markers?.find(marker => marker.id === value);
+      this.onMarkerSelected(selectedMarker);
+    } else if (!!this.selectedMarker){
+      this.onMarkerUnselected();
+    }
+  }
+
+  get selectedMarkerId(): string {
+    return this._selectedMarker?.id;
   }
 
   ngOnInit(): void {
@@ -403,7 +424,8 @@ export class JourneyMapsClientComponent implements OnInit, AfterViewInit, OnDest
   }
 
   /** @internal */
-  onInfoboxCloseClicked(): void {
+  // When a marker has been unselected from outside the map.
+  onMarkerUnselected(): void {
     this.selectedMarker = undefined;
     this.mapService.unselectFeature(this.map);
     this.cd.detectChanges();
@@ -416,7 +438,7 @@ export class JourneyMapsClientComponent implements OnInit, AfterViewInit, OnDest
   }
 
   /** @internal */
-  // When a marker has been selected via search bar.
+  // When a marker has been selected from outside the map.
   onMarkerSelected(marker: Marker): void {
     if (marker?.id !== this.selectedMarker?.id) {
       this.selectedMarker = marker;